refactor(pdfvision): extract focused render and markdown helpers
Keep large coordination files easier to review after recent trainer additions. These helpers preserve behavior while moving self-contained content-box conversion and search-match markdown rendering into nearby focused modules.
